+++ Makefile
@@ -7,19 +7,19 @@ FC?=mpif77
 PAUL?=1
 
 SRCROOT=.
-
 GSLIBDIR=$(GSLIBPATH)
 
 SRCDIR  =$(SRCROOT)/src
-INCDIR  =$(SRCDIR)
 BUILDDIR=$(SRCROOT)/build
 TESTDIR =$(SRCROOT)/tests
 
 TARGET=parRSB
-TESTS=$(TESTDIR)/gmsh/gmsh-test $(TESTDIR)/con/con-test
-LIB=src/lib$(TARGET).a
+TESTS=$(TESTDIR)/con/con-test
+LIB=src/lib$(TARGET).so
 
-INCFLAGS=-I$(INCDIR) -I$(GSLIBDIR)/include
+INCFLAGS=-I$(SRCDIR) -I$(GSLIBDIR)/include
+
+TESTLDFLAGS:=-L$(SRCDIR) -l$(TARGET) -L $(GSLIBDIR)/lib -lgs -lm $(LDFLAGS)
 
 ifneq (,$(strip $(DESTDIR)))
 INSTALL_ROOT = $(DESTDIR)
@@ -27,20 +27,15 @@ else
 INSTALL_ROOT = $(SRCROOT)/build
 endif
 
-TESTLDFLAGS:=-L$(SRCDIR) -l$(TARGET) -L $(GSLIBDIR)/lib -lgs -lm $(LDFLAGS)
 
 # Main source
 CSRCS:= $(SRCDIR)/genmap-vector.c $(SRCDIR)/genmap-algo.c \
 	$(SRCDIR)/genmap-io.c $(SRCDIR)/genmap-comm.c $(SRCDIR)/genmap.c \
 	$(SRCDIR)/parRSB.c $(SRCDIR)/genmap-quality.c \
-	$(SRCDIR)/genmap-chelpers.c $(SRCDIR)/genmap-gmsh.c
-
+	$(SRCDIR)/genmap-chelpers.c
 COBJS:=$(CSRCS:.c=.o)
 
-FSRCS:=
-FOBJS:=$(FSRCS:.f=.o)
-
-SRCOBJS :=$(COBJS) $(FOBJS)
+SRCOBJS:=$(COBJS)
 
 PP= -DGENMAP_LONG_LONG
 ifneq ($(MPI),0)
@@ -86,9 +81,6 @@ endif
 
 $(COBJS): %.o: %.c
 	$(CC) $(CFLAGS) $(PP) $(INCFLAGS) -c $< -o $@
-
-$(FOBJS): %.o: %.f
-	$(FC) $(FFLAGS) $(PP) $(INCFLAGS) -c $< -o $@
 
 .PHONY: tests
 tests: $(TESTS)
