+++ src/genmap-io.h
@@ -1,8 +1,6 @@
 #ifndef _GENMAP_READERS_H_
 #define _GENMAP_READERS_H_
 
-#include "genmap-gmsh.h"
-
 int GenmapCreateHandle_interface(GenmapHandle h);
 int GenmapRead_interface(GenmapHandle h, void *data);
